chore: remove notes on pins
I think these questions are answered with a resounding no!

Throwing together a runtime check was twice as slow as managing
pin state. A lot of the problem is pulling out the specific SquareControl
object we are interested in is slow because of the need to iterate arrays.
This makes the runtime check slower and means using the attack map for
managing pins also not that great.

This could be resolved by a better attack map implementation that supports
more access patterns around directionality.